HC bruto/líq por intervalo: média 10min sem ceil
- hc_bruto = round(mean(cov_phys_10m[i0:i0+3]), 2)
- hc_liq   = round(mean(cov_nr17_10m[i0:i0+3]), 2)
- hc_bruto_max/hc_liq_max no DiaOut: max das médias de 30min (float)
- IntervaloOut.hc_liq e DiaOut.hc_liq_max viram float

Valores fracionários refletem transições dentro do bloco de 30min
(ex: 5.33 = slots 10min (5,5,6)/3). Antes era ceil, que perdia essa
resolução.

Overstaffing agora usa erlang_curve_day().max() como peak_req
(antes derivava de iv.hc_liq, que mudou de semântica).

+++ src/wfm/models.py
@@ -92,7 +92,7 @@ class IntervaloOut:
     horario:      str
     volume:       float
     tmo:          float
-    hc_liq:       int
+    hc_liq:       float
     hc_bruto:     float
     trafico_erl:  float
     fila_pw:      float
@@ -110,7 +110,7 @@ class DiaOut:
     peso_pct:          float
     volume_total:      float
     tmo_medio:         float
-    hc_liq_max:        int
+    hc_liq_max:        float
     hc_bruto_max:      float
     sla_ponderado:     float
     ns_total:          float
